Taks_9

+++ src/main/java/pl/futurecollars/invoicing/model/InvoiceEntry.java
@@ -3,6 +3,7 @@ package pl.futurecollars.invoicing.model;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
+import java.math.BigDecimal;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -13,15 +14,15 @@ public class InvoiceEntry {
     @ApiModelProperty(value = "Product/service description", required = true, example = "Advisory services")
     private String description;
     @ApiModelProperty(value = "Product/service net price", required = true, example = "100")
-    private double price;
+    private BigDecimal price;
     @ApiModelProperty(value = "Product/service VAT amount", required = true, example = "23")
-    private double vatValue;
+    private BigDecimal vatValue;
     @ApiModelProperty(value = "Tax rate", required = true)
     private Vat vatRate;
 
     @JsonCreator
-    public InvoiceEntry(@JsonProperty("description") String description, @JsonProperty("price") double price,
-                        @JsonProperty("vatValue") double vatValue, @JsonProperty("vatRate") Vat vatRate) {
+    public InvoiceEntry(@JsonProperty("description") String description, @JsonProperty("price") BigDecimal price,
+                        @JsonProperty("vatValue") BigDecimal vatValue, @JsonProperty("vatRate") Vat vatRate) {
         this.description = description;
         this.price = price;
         this.vatValue = vatValue;
